requirements.txt
Pillow==10.0.0
ttkthemes==3.2.2
